Extract EntityTitleService interface
