+++ graylog2-server/src/main/java/org/graylog2/bindings/PersistenceServicesBindings.java
@@ -34,6 +34,8 @@ import org.graylog2.notifications.NotificationService;
 import org.graylog2.notifications.NotificationServiceImpl;
 import org.graylog2.rest.resources.entities.preferences.service.EntityListPreferencesService;
 import org.graylog2.rest.resources.entities.preferences.service.EntityListPreferencesServiceImpl;
+import org.graylog2.rest.resources.system.contentpacks.titles.EntityTitleService;
+import org.graylog2.rest.resources.system.contentpacks.titles.EntityTitleServiceImpl;
 import org.graylog2.security.AccessTokenService;
 import org.graylog2.security.AccessTokenServiceImpl;
 import org.graylog2.security.MongoDBSessionService;
@@ -62,5 +64,6 @@ public class PersistenceServicesBindings extends AbstractModule {
         bind(InputStatusService.class).to(MongoInputStatusService.class).asEagerSingleton();
         bind(EntityListPreferencesService.class).to(EntityListPreferencesServiceImpl.class);
         bind(EntitySuggestionService.class).to(MongoEntitySuggestionService.class);
+        bind(EntityTitleService.class).to(EntityTitleServiceImpl.class);
     }
 }
